Update authorization use gate, policy (#12)
* Update authorization use gate, policy

* Create enum, update service

* modify type role migration

+++ app/Services/BaseService.php
@@ -4,6 +4,10 @@ namespace App\Services;
 
 use Illuminate\Contracts\Pagination\Paginator;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Gate;
+use App\Enums\BaseColumn;
+use App\Enums\BaseSort;
+use App\Enums\BaseLimit;
 
 class BaseService
 {
@@ -14,10 +18,10 @@ class BaseService
         $this->model = $model;
     }
 
-    public function getAll($perPage, $query, $filters = [], $columnSearch = null, $termSearch = null): Paginator
+    public function getAll($perPage = BaseLimit::LIMIT_10, $query, $filters = [], $columnSearch = null, $termSearch = null): Paginator
     {
-        $sortColumn = $filters['sortColumn'] ?? 'created_at';
-        $sortOrder = $filters['sortOrder'] ?? 'desc';
+        $sortColumn = $filters['sortColumn'] ?? BaseColumn::COLUMN_CREATED;
+        $sortOrder = $filters['sortOrder'] ?? BaseSort::ORDER_DESC;
 
         $query->orderBy($sortColumn, $sortOrder);
 
@@ -45,6 +49,9 @@ class BaseService
     public function update($id, array $data)
     {
         $model = $this->model->findOrFail($id);
+
+        Gate::authorize('update', $model);
+
         $model->update($data);
 
         return $model;
@@ -53,6 +60,9 @@ class BaseService
     public function delete($id)
     {
         $model = $this->model->findOrFail($id);
+
+        Gate::authorize('delete', $model);
+
         $model->delete();
 
         return $model;
